build-all migrate to NX
