+++ package.json
@@ -20,7 +20,8 @@
     "make-artifacts-package": "ts-node tools/scripts/make-artifacts-package.ts",
     "all:set-timestamp-version": "ts-node tools/scripts/set-timestamp-version.ts",
     "all:update-version": "ts-node tools/scripts/update-version.ts",
-    "all:build": "ts-node tools/scripts/build-all.ts",
+    "all:build": "nx all:build workflows",
+    "all:build:internal": "nx all:build workflows -c internal",
     "all:build-dev": "nx all:build-dev workflows",
     "all:pack-and-copy": "nx run-many -t pack-and-copy",
     "demos:prepare": "nx run devextreme-demos:prepare-js",
